requirements.txt
python-dotenv==1.0.0
fastapi==0.100.0
uvicorn==0.22.0
asyncpg==0.28.0
tortoise-orm==0.19.3